Refactor: domain 패키지 분리
